pyproject.toml
[tool.poetry]
name = "hawc_backend_assignment_project"
version = "0.1.0"
description = ""
authors = ["ShyamSundhar1411 <clashwithchiefrpjyt@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
Django = "^4.0.3"
Pillow = "^9.1.0"
django-countries = {extras = ["pyuca"], version = "^7.3.2"}
django-import-export = "^2.8.0"
django-crispy-forms = "^1.14.0"
django-autoslug = "^1.9.8"
django-filter = "^21.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
